requirements.txt
pdfplumber==0.11.0
python-docx==1.1.2
python-dotenv==1.0.1
google-generativeai==0.8.3
psycopg2-binary==2.9.10
